refactor PostList: separated components
PostList component itself is now differentiated between presentational
PostList component and container component which deals of passing the
redux store's data to the pres' component. The react-redux {connect}
method is the main code in the file.

For now, the data set of posts is defined as initial state. This should
be connected with firebase later on.