init Edit Form rendering logic:
render post edit form when the edit button is clicked. With the help of
state isEditing, the component knows when to show the edit form. While
edit form is mounted and when user access to the other post link, the
isEditing state changes to false again by componentWillReceiveProps
life cycle method.

It's only a wireframe now.