update PostForm: form submit and processing data
In FormSubmit compoent, when the form is submitted, handleSubmit
function is activated and gathers up input value from the form with
refs. Then populate the post object with its values. Then the post
object is dispatched to the reducer.

In the component, instead of using the { dispatch } from redux, to make
presentational component not aware of redux, I used { bindActionCreators
} which binds actions and dispatch method. And it's passed in to
component as a props. So the action method is passed through the child
component.

In the reducer, when the action is passed in, I had to update the
current state. So among the posts exclude the post that I'd like to
update and populate redux store with rest of posts. Then the target post
is just added after. I used filter() method to exclude target object.