add actions for firebase connection:
Actions which network with firebase is created. Those actions can fetch
posts and add post instance. Each action is async action which interacts
with firebase first and then do with redux store.

Right now, the logic is bit rusty but it well communicates with the
server well. Interaction with redux store would be remedied soon.